moved serializers dir inside core
