+++ fedot/core/dag/graph.py
@@ -3,8 +3,8 @@ from typing import TYPE_CHECKING, List, Optional, Union
 from uuid import uuid4
 
 from fedot.core.dag.graph_operator import GraphOperator
+from fedot.core.serializers import GraphSerializer
 from fedot.core.visualisation.graph_viz import GraphVisualiser
-from fedot.serializers import GraphSerializer
 
 if TYPE_CHECKING:
     from fedot.core.dag.graph_node import GraphNode
