+++ fedot/core/operations/operation.py
@@ -4,8 +4,8 @@ from fedot.core.data.data import InputData
 from fedot.core.log import Log, default_log
 from fedot.core.repository.operation_types_repository import OperationMetaInfo
 from fedot.core.repository.tasks import Task, TaskTypesEnum, compatible_task_types
+from fedot.core.serializers import OperationSerializer
 from fedot.core.utils import DEFAULT_PARAMS_STUB
-from fedot.serializers import OperationSerializer
 
 
 class Operation(OperationSerializer):
