+++ fedot/core/serializers/graph_node_serializer.py
