+++ fedot/core/serializers/pipeline_template_serializer.py
